Re-implement node storage
Using our own structs isn't going to work well because we want to be
able to walk the html tree with floki still, regardless of what node
we're talking to.

This simplifies things significantly, we just move around Floki tuples
and then Bauer.Node provides a unified interface for filtering these
tuplies and parsing HTML. Really, it's just a proxy for Floki but adds
some nicities (find vs search).

This also allows us to filter a little more proactively, e.g. this is
bad:

    Page.forms() |> Enum.filter(...)

because we don't want to fetch all of the forms and then traverse the
tree, we just want Floki to handle the heavy stuff, so instead we design
a CSS query with the attributes and use Bauer.Node.find(query).

This commit also starts to define a potential API in the README. I'm not
sold on it, but it's a start